+++ src/api/parameters/asyncConvertContentBodyResponse.ts
@@ -1,0 +1,4 @@
+export interface AsyncConvertContentBodyResponse {
+  /** The asyncId of the macro task to get the converted body. */
+  id: string;
+}