+++ src/api/parameters/getAvailableContentStates.ts
@@ -1,4 +1,7 @@
 export interface GetAvailableContentStates {
-  /** Id of content to get available states for */
-  contentId: string;
+  /**id of content to get available states for*/
+  id?: string;
+
+  /** @deprecated Id of content to get available states for */
+  contentId?: string;
 }