+++ src/api/parameters/getContentState.ts
@@ -1,6 +1,11 @@
 export interface GetContentState {
-  /** The Id of the page whose content state is of interest. */
-  contentId: string;
-  /** Set designation=DRAFT or designation=PUBLISHED. */
-  designation?: string;
+  /** The id of the content whose content state is of interest. */
+  id?: string;
+  /** Set status to one of [current,draft,archived]. */
+  status?: 'current' | 'draft' | 'archived' | string;
+
+  /** @deprecated The Id of the page whose content state is of interest. */
+  contentId?: string;
+  /** @deprecated Set designation=DRAFT or designation=PUBLISHED. */
+  designation?: 'DRAFT' | 'PUBLISHED' | string;
 }