+++ src/api/parameters/getSpaceContentStates.ts
@@ -1,0 +1,4 @@
+export interface GetSpaceContentStates {
+  /** The key of the space to be queried for its content state settings. */
+  spaceKey: string;
+}