+++ src/api/parameters/setContentState.ts
@@ -1,8 +1,16 @@
-export interface SetContentState {
+import { ContentStateRestInput } from '../models';
+
+export interface SetContentState extends ContentStateRestInput {
   /** The Id of the content whose content state is to be set. */
-  contentId: string;
-  /** The Id of the content */
+  id?: string;
+  /**
+   * Status of content onto which state will be placed. If draft, then draft state will change. If current, state will
+   * be placed onto a new version of the content with same body as previous version.
+   */
+  status?: 'current' | 'draft' | string;
+
+  /** @deprecated The Id of the content whose content state is to be set. */
+  contentId?: string;
+  /** @deprecated The Id of the content */
   contentStateId?: number;
-  name?: string;
-  color?: string;
 }